+++ src/cpu/mod.rs
@@ -169,12 +169,12 @@ impl<'a> CPU<'a> {
         self.program_counter = self.read_memory_2_bytes(NMI_ADDRESS);
     }
 
-    pub fn increase_cpu_idle_cycles(&mut self, inc: u16) {
+    pub fn increase_cpu_idle_cycles(&mut self, inc: u8) {
         // if we want to say certain action took x cycles, we just tell the cpu to rest in the next x cycles
         bus_mut!(self).cpu_idle_cycles += inc;
     }
 
-    pub fn decrease_cpu_idle_cycles(&mut self, dec: u16) {
+    pub fn decrease_cpu_idle_cycles(&mut self, dec: u8) {
         // same thing, every cycle we decrease the number of cycles we need to wait
         bus_mut!(self).cpu_idle_cycles -= dec;
     }
